Moving the controller out of the presentation layer
